requirements.txt
streamlit>=1.22.0
openai>=1.5.0
python-dotenv>=1.0.0 